requirements.txt
ccxt
pandas
numpy
requests
setuptools
python-dotenv
